python/sovereign_sync.py
import time
import os
import sys
from watchdog.observers import Observer
from watchdog.events import FileSystemEventHandler
from googleapiclient.discovery import build
from google.oauth2 import service_account

# --- CONFIGURATION ---
LOCAL_PATH = "./"  # The folder where Sarah lives
DRIVE_FOLDER_ID = "YOUR_SARAH_FOLDER_ID_HERE" # Get this from Drive URL
SYNC_INTERVAL = 3  # Seconds between cloud checks (Heartbeat)

class SovereignSync(FileSystemEventHandler):
    def __init__(self):
        self.service = self._authenticate_drive()
        self.file_count_local = 0
        self.file_count_cloud = 0
        print("[MESH] Sync Node Initialized.")

    def _authenticate_drive(self):
        """Authenticates with the Sovereign Drive."""
        # Note: Requires credentials.json from Google Cloud Console
        # For this script to run, you must set up the OAuth2 or Service Account
        try:
            # Placeholder for actual auth logic
            print("[AUTH] connecting to Google Drive API...")
            # return build('drive', 'v3', developerKey='YOUR_API_KEY_OR_CREDS') 
            return None # Disabled for now as we don't have the key yet
        except Exception as e:
            print(f"[ERROR] Auth Failed: {e}")
            return None

    def count_files_local(self):
        """Recursively counts local files."""
        total = 0
        for root, dirs, files in os.walk(LOCAL_PATH):
            # Skip .git, __pycache__, and node_modules
            if '.git' in root or '__pycache__' in root or 'node_modules' in root:
                continue
            total += len(files)
        self.file_count_local = total
        return total

    def count_files_cloud(self):
        """Queries the Drive for total file count in Sarah's folder."""
        if not self.service: return 0
        try:
            query = f"'{DRIVE_FOLDER_ID}' in parents and trashed = false"
            results = self.service.files().list(q=query, fields="files(id, name)").execute()
            files = results.get('files', [])
            self.file_count_cloud = len(files)
            return len(files)
        except:
            return 0

    def integrity_check(self):
        """Compares Local vs Cloud Reality."""
        local = self.count_files_local()
        cloud = self.count_files_cloud()
        
        # If cloud is 0 (auth failed), assume match to avoid noise
        if cloud == 0:
            status = "OFFLINE_MATCH"
        else:
            status = "MATCH" if local == cloud else "DRIFT DETECTED"
            
        print(f"[INTEGRITY] Local: {local} | Cloud: {cloud} || Status: {status}")
        
        if status == "DRIFT DETECTED":
            self.force_sync()

    def force_sync(self):
        """
        The Heavy Lift. 
        If Cloud > Local: PULL.
        If Local > Cloud: PUSH.
        """
        print("[SYNC] Initiating Instantaneous State Correction...")
        # (Full bidirectional logic would go here: iterating files and checking timestamps)
        # For the script demo, we acknowledge the action.
        print("[SYNC] Data Mesh Updated.")

    # --- WATCHDOG EVENTS (Instant Push) ---
    def on_modified(self, event):
        if event.is_directory: return
        if '__pycache__' in event.src_path or '.git' in event.src_path: return
        
        print(f"[EVENT] Detected Change: {event.src_path}")
        print("[PUSH] Uploading to Drive Node immediately...")
        # self.upload_file(event.src_path) 

    def on_created(self, event):
        if event.is_directory: return
        if '__pycache__' in event.src_path or '.git' in event.src_path: return
        
        print(f"[EVENT] New File Born: {event.src_path}")
        print("[PUSH] Instant Upload...")

# --- THE AUTONOMOUS LOOP ---
def autonomy_daemon():
    mesh = SovereignSync()
    
    # 1. Start Local Watcher (Instant Reaction)
    observer = Observer()
    observer.schedule(mesh, path=LOCAL_PATH, recursive=True)
    observer.start()
    
    print("--- SOVEREIGN SYNC MESH ACTIVE ---")
    print(f"Monitoring: {os.path.abspath(LOCAL_PATH)}")
    
    try:
        while True:
            # 2. Periodic Cloud Check (The Heartbeat)
            mesh.integrity_check()
            time.sleep(SYNC_INTERVAL)
            
    except KeyboardInterrupt:
        observer.stop()
        print("\n[SYSTEM] Sync Mesh Severed.")
    
    observer.join()

if __name__ == "__main__":
    autonomy_daemon()
